+++ src/main/kotlin/uk/co/androidalliance/dpad/input/ActivityLauncher.kt
@@ -1,0 +1,7 @@
+package uk.co.androidalliance.dpad.input
+
+import uk.co.androidalliance.dpad.adb.Intent
+
+interface ActivityLauncher {
+    suspend fun startActivity(intent: Intent)
+}